+++ src/Commands.cpp
@@ -306,6 +306,8 @@ static SeqStrings gCommandNames =
     "CmdConfigurePageGrid\0"
     "CmdToggleEditPDF\0"
     "CmdApplyRedactions\0"
+    "CmdCopyAnnotation\0"
+    "CmdPasteAnnotation\0"
     "CmdNone\0"
     "\0";
 
@@ -600,6 +602,8 @@ static i32 gCommandIds[] = {
     CmdConfigurePageGrid,
     CmdToggleEditPDF,
     CmdApplyRedactions,
+    CmdCopyAnnotation,
+    CmdPasteAnnotation,
     CmdNone,
 };
 
@@ -894,6 +898,8 @@ SeqStrings gCommandDescriptions =
     "Configure Page Grid...\0"
     "Toggle Edit PDF\0"
     "Apply Redactions\0"
+    "Copy Annotation\0"
+    "Paste Annotation\0"
     "Do nothing\0"
     "\0";
 // clang-format on